navigation 정리
