+++ src/statics/constants/ScreenName.ts
@@ -1,0 +1,7 @@
+export enum ScreenName {
+	BottomTabNavigator = 'BottomTabNavigator',
+	Home = 'Home',
+	Login = 'Login',
+	Join = 'Join',
+	CreateRecord = 'CreateRecord',
+}
